+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `StorageValue` implementation for `bool`. (#385)
 
 ### Changed
-- `Transaction` `execute` method now returns `u8` instead of `()`. (#385)
+- `Transaction` `execute` method now takes `ExecutionContext`.
+- `Transaction` `execute` method now can have several logical execution statuses representing by `TransactionExecutionStatus`. (#385)
 
 ## 0.4 - 2017-12-08